Increase wxWebRequestCURL's ref count while in progress
When a transfer with a wxWebRequestCURL object is started, a pointer to
that object is stored in its CURL easy handle. That pointer is needed
when then the transfer is complete. Consequently increase the object’s
ref count while it is held by the CURL handle to ensure that it is not
deleted before the transfer is complete.

Correspondingly, remove the pointer and decrease the ref count when the
transfer is complete.