Change cancel method for wxWebRequestCURL objects
Previously wxWebRequestCURL objects were canceled by removing their
CURL easy handle from the CURLM multihandle. Unfortunately the really
only pauses the connection and does not truly cancel it. libcurl
currently doesn’t offer a method for directly canceling a transfer,
but does provide several indirect mechanisms for doing so. This commit
uses three of these mechanisms:

1) It returns a value curl is not expecting from the write callback.
2) It returns a value curl is not expecting from the read callback.
3) It returns a nonzero value from the progress callback.

As soon is one of these 3 things happens, curl will stop the transfer.

This commit also adds the progress callback so that it can be used for
cancellation. Making this addition presents a few complications because
over the years curl has changed the signature of the callback and added
a special return value besides zero that can be used to continue the
transfer. A combination of compile-time and run-time checks is used to
make sure the appropriate callback and preferred return value are used.