Add a progress callback for wxWebRequestCURL
This commit adds a progress callback for use with wxWebRequestCURL
objects. This has several complications because over the years curl has
changed the signature of the callback and added a special return value
that can be used to continue the transfer. Prior to that addition, the
callback was supposed to return zero to continue the transfer.

A combination of compile-time and run-time checks is used to make sure
the appropriate callback and preferred return value are used.