Record the active socket in wxWebRequestCURL
When canceling or failing a transfer managed by a wxWebRequestCURL
object, we need to know the socket descriptor being used for the
transfer. Unfortunately on some versions of windows with older versions
of the curl library, we can’t get the socket directly from curl.

This commit works around this by storing the active socket descriptor in
the wxWebRequestCURL object using curl’s callbacks for creating and
closing a socket. The active socket is recorded when a socket is created
and set to an invalid socket when the socket is closed.